y in PlayerDiggingPacket is now an unsigned byte

+++ src/main/java/tk/maincraft/util/mcpackets/packet/PlayerDiggingPacket.java
@@ -5,7 +5,7 @@ import tk.maincraft.util.mcpackets.Packet;
 public interface PlayerDiggingPacket extends Packet {
     byte getStatus();
     int getX();
-    byte getY();
+    int getY();
     int getZ();
     byte getFace();
 }
